+++ src/main/java/domain/Application.java
@@ -1,4 +1,4 @@
-package lotto;
+package domain;
 
 import controller.LottoController;
 import view.InputView;
