+++ src/main/java/domain/lotto/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package domain.lotto;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
