+++ src/main/java/domain/lotto/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package domain.lotto;
 
 import constants.ErrorMessage;
 import java.util.ArrayList;
