+++ src/test/java/lotto/BonusNumberTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import constants.ErrorMessage;
+import domain.winningLotto.BonusNumber;
+import domain.lotto.LottoNumber;
+import domain.winningLotto.WinningNumbers;
 import java.util.List;
 import org.junit.jupiter.api.Test;
