+++ src/test/java/lotto/LottoGeneratorTest.java
@@ -2,6 +2,9 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.lotto.Lotto;
+import domain.lotto.LottoGenerator;
+import domain.lotto.LottoNumber;
 import java.util.List;
 import org.junit.jupiter.api.Test;
