+++ src/test/java/lotto/LottoValidatorTest.java
@@ -3,6 +3,8 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import constants.ErrorMessage;
+import domain.lotto.Lotto;
+import domain.lotto.LottoNumber;
 import java.util.List;
 import org.junit.jupiter.api.Test;
