+++ src/test/java/lotto/RevenueCalculatorTest.java
@@ -2,6 +2,9 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.lotto.PurchaseAmount;
+import domain.statistics.RevenueCalculator;
+import domain.statistics.WinningRank;
 import java.util.Arrays;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
