+++ src/test/java/lotto/WinningLottoTest.java
@@ -2,6 +2,10 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.winningLotto.BonusNumber;
+import domain.lotto.LottoNumber;
+import domain.winningLotto.WinningLotto;
+import domain.winningLotto.WinningNumbers;
 import java.util.List;
 import org.junit.jupiter.api.Test;
